top: provide for disabling of tty output optimizations
This is just for timings to prove how expensive output
operations really are when no optimizations were used.

Over 20 years ago, when top introduced optimization of
tty output, the speedup realized made him 3 to 6 times
faster than the previous 3.1.0 version. And, while the
improvement is not nearly as great today, the approach
to minimizing tty outputs still seems worth retaining.

Signed-off-by: Jim Warner <james.warner@comcast.net>