top: make immune to order of settings in config_wins()
If a config variable setting is moved after its public
release, without any changes top could no longer honor
the older configuration files. So, we must modify that
original code devoted to parsing of variable settings.

The following two approaches could solve this problem:

1) Separate 'if' statements with fscanf to accommodate
the change in settings. Which 'if' is executed depends
on that RCF_VERSION_ID being less than the latest 'n'.

That approach is ugly, inflexible, yet most efficient.

2) Introduce a table driven approach and treat those 2
affected lines as a single glob searched for settings.

This is prettier, more flexible, but not as efficient.

So, since I cannot decide which approach to take, I've
implemented both then defaulted to the more efficient.

Signed-off-by: Jim Warner <james.warner@comcast.net>