library: implement our own 'stable' mergesort function
Thanks to Carl Thompson, and the issue cited below, we
will now be able to guarantee a stable sort with those
APIs which export such a function. This will currently
benefit the ps program that allows multiple sort keys.

And, while we would have had a stable sort with glibc,
as Carl pointed out not all distros used that library.

Reference(s):
https://gitlab.com/procps-ng/procps/-/issues/398

Signed-off-by: Jim Warner <james.warner@comcast.net>