library: 4 modules exploit 'stable' mergesort function
Along with that sort function name change, this commit
now honors our documentation regarding a return value.

A 0 return from that new mergesort_r function can only
mean it suffered ENOMEM with a malloc call. In the man
pages we are told a function returning an address will
indicate failure through a NULL pointer and the reason
will then be reflected in the traditional errno value.

Signed-off-by: Jim Warner <james.warner@comcast.net>