misc: remove that stablesort routine in favor of qsort
When Roman Zilka refactored the signals logic, in that
commit referenced below, the 'stablesort' function was
also added to strutils.c. The reason provided was that
stability of the stdlib qsort routine was unspecified.

But, the stability issue is a concern only if multiple
sort keys are used and that's never true in signals.c.

So, this patch will trade use of stablesort for qsort.

Reference(s):
. Jan, 2023 - update to signals
commit ac89d63c5a43ab1b7443a0ee7b49baacc1c037a0

Signed-off-by: Jim Warner <james.warner@comcast.net>