library: squeeze a tad more performance out of sorting
When I offered our project leader the patch referenced
below, I said there was a 90% certainty there would be
no more for the sort module. Here's the remaining 10%.

This patch will trade much pointer arithmetic that was
frequently repeated for a succinct postfix expression.

For an optimized compile using -O3, these changes will
reduce module size by 64 bytes for a new total of 992.
Plus 12 iterated instructions have been eliminated out
of a total of 243, reducing total instructions to 231.

And, while the above improvements are not likely to be
measurable, differences between our sort and the glibc
implementation most certainly are. For example, if one
limits sort to a constant number of tasks (i.e. 1500),
then these are some results in millionths of a second:

   total key(s)   glibc qsort    our mergesort
   ------------   ------------   -------------
       1            0.000060       0.000054
       2            0.000160       0.000121
       6            0.000592       0.000549
       8            0.000711       0.000631
       11           0.001239       0.000959

[ total key(s) represents calls to the sort function ]
[ and, where more than one, emphasize the importance ]
[ the 'stability' factor. Time measured is that of a ]
[ sort function's overhead only, not any other cost. ]

Lastly, and unrelated to those code changes above, I'm
a little surprised to find that some identifiers which
were changed with the patch below, were changed AGAIN!

Oh well, the new names are definately an improvement!!

Reference(s):
, Nov, 2025 - refactor identifiers
commit db4a4c561d40f8c78df9cb1fe6f5dcc2a0155c55

Signed-off-by: Jim Warner <james.warner@comcast.net>