Fix integer overflow in scale_size() for large memory values
Under some circumstances, you may get a very large swapfree value.
This will be reported as a negative value as it overflows the
variable handling it. The scale_size() accepts unsigned values
but uses signed values inside it, giving the unusual result.

This commit is largely @Fgaoxing work on !275 but without all
the whitespace noisy change to it.

References:
 #397
 !275

Signed-off-by: Craig Small <csmall@dropbear.xyz>