docs: w ignores --ip-addr on systemd platforms
On platforms that use libsystemd to track users, we use
sd_session_get_remote_host() to find the "from" field for a session.
There is no sd_session_get_remote_addr() to directly get the IP
address of the remote session. Using a DNS resolution, such as
via getaddrinfo() is problematic because it's up to how DNS
is setup and that could be wrong, or provide multiple answers, or both.

sd_session_get_remote_host() may return a hostname or IP address,
I'm not exactly sure why it decides to do either.

References:
 #396

Signed-off-by: Craig Small <csmall@dropbear.xyz>