ps: fix segfault under selinux without libnuma
Currently, running `ps auxZ` with SELinux enabled without libnuma
installed throws a segfault at the ps_freecon call. This is because a
previous dlopen() call failed and the dlerror was not reset since,
causing the code to think the dlsym call to get freecon failed.

Fix this by resetting dlerror before the first dlsym call. Additionally,
add a check that ps_freecon is non-null before entering the libselinux
path.