library: reduce function call overhead for utf-8 logic
The utf-8 validation code found in the escape.c module
is structured so the u8charlen function was called for
every 'character' (ascii or multi-byte) in the string.

This commit just transforms the code in u8charlen into
a loop so that it will be called just once per string.

[ when tested with top, using a consistent number of ]
[ tasks and rcfile specifying all fields requiring a ]
[ call to escape_str, performance improved about 6%. ]

Reference(s):
May, 2024 - escape refactored, u8charlen introduced
commit 73d332e8f174d51ddf44a14573bac623601e8c00

Signed-off-by: Jim Warner <james.warner@comcast.net>