w: fix help text for --from command option
There is a W_SHOWFROM build option that toggles showing or hiding the
"FROM" field by default. However, the usage text incorrectly states that
the "-f/--from" command option always enables this field, while in
reality it toggles the field on or off depending on the W_SHOWFROM build
option.

Fix the help text for the "-f/--from" command option to make the usage
logically correct.

Fixes: 7af14af6 ("w: new usage & fix coding style")
Signed-off-by: Oleksandr Suvorov <cryosay@gmail.com>